refactor: 뷰로 분리
#76
